requirements.txt
streamlit
marqo
google-generativeai
pillow
requests